+++ test/suite/functional/Generator/Schema/ItemPhp70.php
@@ -56,6 +56,10 @@ class Item implements SerializableInterface, JsonSerializable
     /** @var ItemMandatoryObject */
     private $mandatoryObject;
 
+    private $mandatoryMixed;
+
+    private $mandatoryAnyOf;
+
     /** @var ItemNullableObject|null */
     private $nullableObject;
 
@@ -82,6 +86,9 @@ class Item implements SerializableInterface, JsonSerializable
 
     /** @var string[]|null */
     private $optionalArray;
+
+    /** @var mixed[]|null */
+    private $optionalMixedArray;
 
     /** @var string[]|null */
     private $optionalArrayWithMinMaxItems;
@@ -114,7 +121,7 @@ class Item implements SerializableInterface, JsonSerializable
      *
      * @throws RequestValidationException
      */
-    public function __construct(int $mandatoryInteger, string $mandatoryString, string $mandatoryEnum, DateTimeInterface $mandatoryDate, $mandatoryNullableDate, float $mandatoryFloat, bool $mandatoryBoolean, array $mandatoryArray, array $mandatoryArrayWithMinItems, ItemMandatoryObject $mandatoryObject)
+    public function __construct(int $mandatoryInteger, string $mandatoryString, string $mandatoryEnum, DateTimeInterface $mandatoryDate, $mandatoryNullableDate, float $mandatoryFloat, bool $mandatoryBoolean, array $mandatoryArray, array $mandatoryArrayWithMinItems, ItemMandatoryObject $mandatoryObject, $mandatoryMixed, $mandatoryAnyOf)
     {
         $this->mandatoryInteger = $mandatoryInteger;
         $this->mandatoryString  = $mandatoryString;
@@ -132,6 +139,8 @@ class Item implements SerializableInterface, JsonSerializable
         }
         $this->mandatoryArrayWithMinItems = $mandatoryArrayWithMinItems;
         $this->mandatoryObject            = $mandatoryObject;
+        $this->mandatoryMixed             = $mandatoryMixed;
+        $this->mandatoryAnyOf             = $mandatoryAnyOf;
     }
 
     /**
@@ -208,6 +217,16 @@ class Item implements SerializableInterface, JsonSerializable
     public function setOptionalArray(array $optionalArray): self
     {
         $this->optionalArray = $optionalArray;
+
+        return $this;
+    }
+
+    /**
+     * @param mixed[] $optionalMixedArray
+     */
+    public function setOptionalMixedArray(array $optionalMixedArray): self
+    {
+        $this->optionalMixedArray = $optionalMixedArray;
 
         return $this;
     }
@@ -389,6 +408,16 @@ class Item implements SerializableInterface, JsonSerializable
         return $this->mandatoryObject;
     }
 
+    public function getMandatoryMixed()
+    {
+        return $this->mandatoryMixed;
+    }
+
+    public function getMandatoryAnyOf()
+    {
+        return $this->mandatoryAnyOf;
+    }
+
     /**
      * @return ItemNullableObject|null
      */
@@ -459,6 +488,14 @@ class Item implements SerializableInterface, JsonSerializable
     public function getOptionalArray()
     {
         return $this->optionalArray;
+    }
+
+    /**
+     * @return mixed[]|null
+     */
+    public function getOptionalMixedArray()
+    {
+        return $this->optionalMixedArray;
     }
 
     /**
@@ -538,6 +575,8 @@ class Item implements SerializableInterface, JsonSerializable
         $fields['mandatoryArray']             = $this->mandatoryArray;
         $fields['mandatoryArrayWithMinItems'] = $this->mandatoryArrayWithMinItems;
         $fields['mandatoryObject']            = $this->mandatoryObject->toArray();
+        $fields['mandatoryMixed']             = $this->mandatoryMixed;
+        $fields['mandatoryAnyOf']             = $this->mandatoryAnyOf;
         $fields['nullableObject']             = $this->nullableObject !== null ? $this->nullableObject->toArray() : null;
         $fields['nullableDate']               = $this->nullableDate   !== null ? $this->nullableDate->format(DATE_RFC3339) : null;
         if ($this->optionalInteger !== null) {
@@ -560,6 +599,9 @@ class Item implements SerializableInterface, JsonSerializable
         }
         if ($this->optionalArray !== null) {
             $fields['optionalArray'] = $this->optionalArray;
+        }
+        if ($this->optionalMixedArray !== null) {
+            $fields['optionalMixedArray'] = $this->optionalMixedArray;
         }
         if ($this->optionalArrayWithMinMaxItems !== null) {
             $fields['optionalArrayWithMinMaxItems'] = $this->optionalArrayWithMinMaxItems;