+++ test/suite/functional/Generator/Schema/ItemPhp72.php
@@ -56,6 +56,10 @@ class Item implements SerializableInterface, JsonSerializable
     /** @var ItemMandatoryObject */
     private $mandatoryObject;
 
+    private $mandatoryMixed;
+
+    private $mandatoryAnyOf;
+
     /** @var ItemNullableObject|null */
     private $nullableObject;
 
@@ -82,6 +86,9 @@ class Item implements SerializableInterface, JsonSerializable
 
     /** @var string[]|null */
     private $optionalArray;
+
+    /** @var mixed[]|null */
+    private $optionalMixedArray;
 
     /** @var string[]|null */
     private $optionalArrayWithMinMaxItems;
@@ -113,7 +120,7 @@ class Item implements SerializableInterface, JsonSerializable
      *
      * @throws RequestValidationException
      */
-    public function __construct(int $mandatoryInteger, string $mandatoryString, string $mandatoryEnum, DateTimeInterface $mandatoryDate, ?DateTimeInterface $mandatoryNullableDate, float $mandatoryFloat, bool $mandatoryBoolean, array $mandatoryArray, array $mandatoryArrayWithMinItems, ItemMandatoryObject $mandatoryObject)
+    public function __construct(int $mandatoryInteger, string $mandatoryString, string $mandatoryEnum, DateTimeInterface $mandatoryDate, ?DateTimeInterface $mandatoryNullableDate, float $mandatoryFloat, bool $mandatoryBoolean, array $mandatoryArray, array $mandatoryArrayWithMinItems, ItemMandatoryObject $mandatoryObject, $mandatoryMixed, $mandatoryAnyOf)
     {
         $this->mandatoryInteger = $mandatoryInteger;
         $this->mandatoryString  = $mandatoryString;
@@ -131,6 +138,8 @@ class Item implements SerializableInterface, JsonSerializable
         }
         $this->mandatoryArrayWithMinItems = $mandatoryArrayWithMinItems;
         $this->mandatoryObject            = $mandatoryObject;
+        $this->mandatoryMixed             = $mandatoryMixed;
+        $this->mandatoryAnyOf             = $mandatoryAnyOf;
     }
 
     public function setNullableObject(?ItemNullableObject $nullableObject): self
@@ -201,6 +210,16 @@ class Item implements SerializableInterface, JsonSerializable
     public function setOptionalArray(array $optionalArray): self
     {
         $this->optionalArray = $optionalArray;
+
+        return $this;
+    }
+
+    /**
+     * @param mixed[] $optionalMixedArray
+     */
+    public function setOptionalMixedArray(array $optionalMixedArray): self
+    {
+        $this->optionalMixedArray = $optionalMixedArray;
 
         return $this;
     }
@@ -379,6 +398,16 @@ class Item implements SerializableInterface, JsonSerializable
         return $this->mandatoryObject;
     }
 
+    public function getMandatoryMixed()
+    {
+        return $this->mandatoryMixed;
+    }
+
+    public function getMandatoryAnyOf()
+    {
+        return $this->mandatoryAnyOf;
+    }
+
     public function getNullableObject(): ?ItemNullableObject
     {
         return $this->nullableObject;
@@ -425,6 +454,14 @@ class Item implements SerializableInterface, JsonSerializable
     public function getOptionalArray(): ?array
     {
         return $this->optionalArray;
+    }
+
+    /**
+     * @return mixed[]|null
+     */
+    public function getOptionalMixedArray(): ?array
+    {
+        return $this->optionalMixedArray;
     }
 
     /**
@@ -483,6 +520,8 @@ class Item implements SerializableInterface, JsonSerializable
         $fields['mandatoryArray']             = $this->mandatoryArray;
         $fields['mandatoryArrayWithMinItems'] = $this->mandatoryArrayWithMinItems;
         $fields['mandatoryObject']            = $this->mandatoryObject->toArray();
+        $fields['mandatoryMixed']             = $this->mandatoryMixed;
+        $fields['mandatoryAnyOf']             = $this->mandatoryAnyOf;
         $fields['nullableObject']             = $this->nullableObject !== null ? $this->nullableObject->toArray() : null;
         $fields['nullableDate']               = $this->nullableDate   !== null ? $this->nullableDate->format(DATE_RFC3339) : null;
         if ($this->optionalInteger !== null) {
@@ -505,6 +544,9 @@ class Item implements SerializableInterface, JsonSerializable
         }
         if ($this->optionalArray !== null) {
             $fields['optionalArray'] = $this->optionalArray;
+        }
+        if ($this->optionalMixedArray !== null) {
+            $fields['optionalMixedArray'] = $this->optionalMixedArray;
         }
         if ($this->optionalArrayWithMinMaxItems !== null) {
             $fields['optionalArrayWithMinMaxItems'] = $this->optionalArrayWithMinMaxItems;