+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [5.3.0] - 2021-04-13
+### Added
+ - Support for `mixed` parameter type
+ - Support for anyOf `mixed` parameter type
+
+### Fixed
+ - `mixed` arrays rendered incorrectly
+ - Wrong typehint in case of nullable arrays
+
 ## [5.2.0] - 2021-04-06
 ### Added
  - Support for `minimum`, `maximum`, `exclusiveMinimum`, `exclusiveMaximum`, `minItems`, `maxItems`, `pattern`, `maxLength`, `minLength` validations